+++ glide-core/tests/test_cluster_client.rs
@@ -884,7 +884,7 @@ mod cluster_client_tests {
                 .expect("build cluster client");
 
             let mut conn: redis::cluster_async::ClusterConnection = cluster_client
-                .get_async_connection(None, None)
+                .get_async_connection(None, None, None)
                 .await
                 .expect("connect to cluster");